// src/application/dtos/user/userUpdate.dto.ts
export class UserUpdateDto {
  id: number = 0;
  address: string = "";
  age: string = "";
  birthday: string = "";
  cellPhone: string = "";
  documentNumber: string = "";
  email: string = "";
  idDocumentsType: number = 0;
  idRole: number = 0;
  lastName: string = "";
  names: string = "";
  password: string = "";
  repeatPassword: string = "";
  state: string = "";
  userUpdate: number = 0;
}